backend/get-started-open-feature/client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-appconfiguration-provider>=1.3.0",
    "azure-appconfiguration>=1.7.1",
    "azure-identity>=1.19.0",
    "featuremanagement>=1.0.0",
    "launchdarkly-openfeature-server>=0.3.0",
    "openfeature-sdk>=0.7.4",
    "ruff>=0.8.2",
]
